+++ src/Integration.DevKit.TaskMgmt/Implementations/TimeStrategy_Hourly.cs
@@ -31,7 +31,7 @@ public sealed class TimeStrategy_Hourly : Time_IterationStrategy
     /// By adding an hour to the <c>Target</c> time rather than the <c>Current</c> time, this strategy 
     /// maintains a consistent schedule even if the task work takes several minutes or hours to complete.
     /// </remarks>
-    protected override DateTime ComputeNextTargetDTM(int currentIteration)
+    protected override DateTime ComputeNextTargetDTM(long currentIteration)
     {
         return LastTargetDTM.AddHours(1);
     }
